Move Component analysis/transformer to own file
